Create multiplatform payment sheet composable
