+++ app-shared/build.gradle.kts
@@ -69,9 +69,6 @@ kotlin {
                 implementation(libs.kotlinx.dateTime)
                 implementation(libs.kotlinx.coroutines)
                 implementation(libs.koin.core)
-
-                implementation(libs.stripe.android)
-                implementation(libs.stripe.connections)
             }
         }
         val commonTest by getting {
@@ -90,6 +87,13 @@ kotlin {
                 implementation(libs.ktor.client.okhttp)
 
                 implementation(compose.desktop.currentOs)
+            }
+        }
+
+        val androidMain by getting {
+            dependencies {
+                implementation(libs.stripe.android)
+                implementation(libs.stripe.connections)
             }
         }
     }
