--- app-shared/src/commonMain/kotlin/ModuleInfo.kt
@@ -1,7 +1,0 @@
-/*
- * Copyright (c) 2024 Pointyware. Use of this software is governed by the GPL-3.0 license.
- */
-
-package org.pointyware.xyz.shared
-
-const val MODULE_NAME = ":app-shared"
