+++ app-shared/src/commonMain/kotlin/navigation/PaymentsRouting.kt
@@ -1,34 +1,16 @@
 package org.pointyware.xyz.shared.navigation
 
+import androidx.compose.runtime.Composable
 import org.pointyware.xyz.core.navigation.XyzRootScope
 import org.pointyware.xyz.core.navigation.toTypedKey
-import org.pointyware.xyz.feature.login.navigation.paymentsRoute
+import org.pointyware.xyz.ui.PaymentSheet
 
 val paymentsRoute = "payment-sheet".toTypedKey<Unit>()
 
+@Composable
 fun XyzRootScope.payments() {
 
     location(paymentsRoute) {
-
-        val paymentSheet = rememberPaymentSheet(
-            createIntentCallback = { paymentMethod, shouldSavePaymentMethod ->
-
-                val clientSecret = "returned-from-our-server"
-                CreateIntentResult.Success(clientSecret)
-            },
-            paymentResultCallback = { result ->
-                when (result) {
-                    PaymentSheetResult.Canceled -> {
-                        TODO("Handle cancellation")
-                    }
-                    PaymentSheetResult.Completed -> {
-                        TODO("Handle completion")
-                    }
-                    is PaymentSheetResult.Failed -> {
-                        result.error.printStackTrace()
-                    }
-                }
-            }
-        )
+        PaymentSheet()
     }
 }
