+++ app-shared/src/commonMain/kotlin/org/pointyware/xyz/ui/PaymentSheet.kt
@@ -1,0 +1,6 @@
+package org.pointyware.xyz.ui
+
+import androidx.compose.runtime.Composable
+
+@Composable
+expect fun PaymentSheet()
