+++ app-shared/src/iosMain/kotlin/org/pointyware/xyz/ui/PaymentSheet.ios.kt
@@ -1,0 +1,10 @@
+package org.pointyware.xyz.ui
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+
+@Composable
+actual fun PaymentSheet() {
+    Text("PaymentSheet Not Yet Supported on iOS")
+}
