+++ app-shared/src/jvmMain/kotlin/org/pointyware/xyz/ui/PaymentSheet.jvm.kt
@@ -1,0 +1,9 @@
+package org.pointyware.xyz.ui
+
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+
+@Composable
+actual fun PaymentSheet() {
+    Text("PaymentSheet Not Yet Supported on JVM")
+}
